requirements.txt
requests==2.7.0
elasticsearch==1.5.0
cython==0.22
thriftpy==0.3.1
pyramid==1.5.7
dogpile.cache==0.5.6
pylibmc==1.5.0
-e git+https://github.com/scieloorg/xylose@0.21#egg=xylose
-e git+https://github.com/scieloorg/access_stats@0.1.2#egg=publication_stats
-e git+https://github.com/scieloorg/thriftpy-wrap@0.1.1#egg=thriftpywrap
